Create instance of `PatternFieldGenerator` for each field
